Reservation request ownership
Interposed logic of purposeful request manipulation by the author or offeror to whom the request was submitted.

+++ src/requests/requests.service.ts
@@ -2,6 +2,7 @@ import {
   ConflictException,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { RequestForReservationDTO } from './dto/request-for-reservartion.dto';
@@ -9,6 +10,7 @@ import { RequestsRepository } from './requests.repository';
 import { Request } from './request.entity';
 import { GetRequestsFilterDTO } from './dto/get-requests-filter.dto';
 import { ReservationsService } from 'src/reservations/reservations.service';
+import { Account } from 'src/auth/account.entity';
 
 @Injectable()
 export class RequestsService {
@@ -24,21 +26,30 @@ export class RequestsService {
     return this.requestsRepository.insertRequest(requestForReservationDTO);
   }
 
-  getRequests(getRequestsFilterDTO: GetRequestsFilterDTO): Promise<Request[]> {
-    return this.requestsRepository.selectRequests(getRequestsFilterDTO);
+  getRequests(
+    account: Account,
+    getRequestsFilterDTO: GetRequestsFilterDTO,
+  ): Promise<Request[]> {
+    return this.requestsRepository.selectRequests(
+      account,
+      getRequestsFilterDTO,
+    );
   }
 
-  getRequest(idRequests: string): Promise<Request> {
-    return this.requestsRepository.selectRequest(idRequests);
+  async getRequestById(account: Account, idRequests: string): Promise<Request> {
+    return this.requestsRepository.selectRequest(account, idRequests);
   }
 
-  async retreatRequest(idRequests: string): Promise<void> {
-    // if requests was not found
+  async retreatRequest(account: Account, idRequests: string): Promise<void> {
+    const request = await this.requestsRepository.findOne({ idRequests });
+    // if request was not found
     if (!(await this.requestsRepository.findOne(idRequests)))
       throw new NotFoundException('Request was not found.');
+    // if request hasn't been confirmed as reservation
     if (this.reservationsService.getReservation(idRequests))
-      // if request hasn't been confirmed as reservation
       throw new ConflictException('Request is already confirmed.');
-    return this.requestsRepository.deleteRequest(idRequests);
+    // if request is owned by an account
+    if (request.offeree.account == account)
+      return this.requestsRepository.deleteRequest(idRequests);
   }
 }
